Replace base image with upstream Silverblue
Reason to switch is, that the intended image use case does not
care about most modifications of the `ublue-main` image. Instead,
some things need to be disabled as they cause problems in
the intended environment.

New base image:
  - https://gitlab.com/fedora/ostree/ci-test
  - https://quay.io/organization/fedora-ostree-desktops

Image used by the `ublue` project as well:
https://github.com/ublue-os/main/blob/3eae314e16c1516b3e550006bb7517d34bdc49e5/Containerfile

It might be possible to switch to
https://quay.io/repository/fedora/fedora-silverblue at some point, but
seems currently not recommended.

Add new Ansible role for configuration of system updates via
`rpm-ostreed` and `flatpak`, which is missing in the new base image.